LayoutTests/crypto/crypto-random-values-limits-expected.txt

Tests the limits of crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'crypto' in window is true
PASS 'getRandomValues' in window.crypto is true
PASS crypto.getRandomValues(almostTooLargeArray) did not throw exception.
PASS crypto.getRandomValues(tooLargeArray) threw exception QuotaExceededError (DOM Exception 22): The quota has been exceeded..
PASS successfullyParsed is true

TEST COMPLETE
